+++ RELEASE-NOTES.md
@@ -15,6 +15,12 @@ Add refunds functionality
 ### Version 4.9.16-wcag20-exui-4"
 **FPLA: Accessibility Issues Found in ExUI** FPLA: Accessibility Issues Found in ExUI
 
+### Version 4.12.16-task-event-completion"
+**Task event completion and state machine** Task event completion and state machine
+
+### Version 4.12.6-limit-search-results-10k-400-5"
+**Limit search results to 10k / 400 pages** Limit search results to 10k / 400 pages
+
 ### Version 4.9.15-state-machine"
 **Task event completion and state machine** Task event completion and state machine